requirements.txt
streamlit>=1.30.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
aiohttp>=3.9.0
feedparser>=6.0.0
wikipedia>=1.4.0
beautifulsoup4>=4.12.0
streamlit-option-menu>=0.3.6
python-dotenv>=1.0.0
pydeck>=0.8.0
